Removes Kakao channel coupon feature
This commit removes the Kakao channel friend coupon feature and related code.

The feature is removed from the top banners, login event slider and other related components to streamline the user experience and focus on core functionalities. Also removed kakao related code and images that are no longer needed, as well as the kakao button components.

This change simplifies the application and removes outdated promotional content.